move file to cli
