+++ aqueduct/lib/src/cli/commands/db_upgrade.dart
@@ -7,7 +7,7 @@ import 'package:aqueduct/src/cli/mixins/project.dart';
 import 'package:aqueduct/src/cli/scripts/run_upgrade.dart';
 import 'package:aqueduct/src/db/postgresql/postgresql_persistent_store.dart';
 import 'package:aqueduct/src/db/query/query.dart';
-import 'package:aqueduct/src/db/schema/migration_source.dart';
+import 'package:aqueduct/src/cli/migration_source.dart';
 import 'package:aqueduct/src/db/schema/schema.dart';
 import 'package:isolate_executor/isolate_executor.dart';
