+++ aqueduct/test/command/migration_execution_test.dart
@@ -5,7 +5,7 @@ import 'dart:io';
 
 import 'package:aqueduct/aqueduct.dart';
 import 'package:aqueduct/src/compilers/project_analyzer.dart';
-import 'package:aqueduct/src/db/schema/migration_source.dart';
+import 'package:aqueduct/src/cli/migration_source.dart';
 import 'package:test/test.dart';
 
 import 'cli_helpers.dart';
